fix(rsvp): address review nits
